Clean up address handling.
This is actually a bit tricky, since proper functioning of the program
was actually dependent on the address parsing in nlsocket.cpp being
buggy.

The problem was that scopeid link or host addresses would not be
properly parsed -- in fact, only global addresses would be parsed.

Then in radv6.cpp, none of these addresses would actually be used,
since it wanted link-scoped addresses.

Effectively, because of default construction of the asio address
structure, the any() [all zeros] address would be used.

This was actually correct!  So the program functioned fine with these
bugs in place, aside from printing errors about unrecognized scopeid
types in nlsocket.cpp.

The reason this was correct is itself subtle.  The kernel is
filling in the source address on the actual packets that we are
sending, and we have bound them to interface by other calls in the
multicast or raw socket interfaces.  Any socket in the correct
family would work, so long as it was NOT a link-local socket,
which will fail the bind() call.

Thus, just explicitly use any() addresses where appropriate, and
then the bugs in nlsocket.cpp can be fixed.

A further small issue is that scopeid was being passed in
address construction in nlsocket where the ifindex is actually
correct.  The ASIO documentation is confusing here, but strace
doesn't lie.

All these changes taken together have no functional effect, but
make the code much easier to understand and more importantly, more
correct.